% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_mn_logo.R
\name{add_mn_logo}
\alias{add_mn_logo}
\title{Add MN state logo to ggplot}
\usage{
add_mn_logo(
  plot = NULL,
  type = "horizontal",
  logo_width = 0.2,
  logo_path = NULL,
  bg_color = NULL
)
}
\arguments{
\item{plot}{A ggplot object. Must be provided.}

\item{type}{Character. Which logo to use. Options are "primary" (default) or "primary-inverse".}

\item{logo_width}{Numeric. Width of the logo relative to the plot (default 0.2 = 20\%).}

\item{logo_path}{Character. Optional full path or URL to a custom logo. If NULL, uses the MN logo based on type.
#' @param bg_color Character or NULL. Background color of entire plot. Default is transparent / white.}
}
\value{
A ggplot object with the MN logo image added.
}
\description{
Adds a Minnesota state agency or department logo to the bottom-right of a ggplot object.
}
\examples{
\dontrun{
p <- ggplot(mtcars, aes(mpg, wt)) + geom_point()
p |> add_mn_logo(logo_width = 0.20)
}
}
\keyword{logo}
