% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_theme.R
\name{base_theme}
\alias{base_theme}
\title{Base ggplot2 theme for internal package use}
\usage{
base_theme(
  base_family = "TT Arial",
  base_size = 14,
  plot_title_family = base_family,
  plot_title_size = 22,
  plot_title_face = "bold",
  plot_title_margin = 11,
  subtitle_family = base_family,
  subtitle_size = 14,
  subtitle_face = "plain",
  subtitle_margin = 15,
  legend_size = 15,
  strip_text_family = base_family,
  strip_text_size = 14,
  strip_text_face = "plain",
  caption_family = base_family,
  caption_size = 12,
  caption_face = "italic",
  caption_margin = 10,
  caption_align = "left",
  axis_text_size = base_size,
  axis_title_family = subtitle_family,
  axis_title_size = 14,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = unit(c(0.2, 0.4, 0.2, 0.2), "cm"),
  grid = TRUE,
  grid_col = "#cccccc",
  axis = FALSE,
  axis_col = "#cccccc",
  ticks = FALSE,
  hide_legend = TRUE,
  hide_legend_title = TRUE
)
}
\arguments{
\item{base_family}{Font family}

\item{base_size}{Base font size}

\item{plot_title_size}{Size of plot titles}

\item{grid}{Logical or character controlling grid lines}

\item{axis}{Logical or character controlling axis lines}

\item{ticks}{Logical; show axis ticks}
}
\description{
Base ggplot2 theme for internal package use
}
\keyword{"Helvetica}
\keyword{Neue",Helvetica,"Helvetica-Narrow",Arial,sans-serif',}
\keyword{internal}
