% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_theme_dark.R
\name{base_theme_dark}
\alias{base_theme_dark}
\title{Fill plot background and change all text colors}
\usage{
base_theme_dark(bg_color = "#003865", text_color = "#F0F0F0")
}
\description{
Fill plot background and change all text colors
}
\keyword{internal}
