% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnthemes-package.R
\docType{package}
\name{mnthemes-package}
\alias{mnthemes}
\alias{mnthemes-package}
\title{mnthemes: Minnesota brand ggplot2 themes and helper functions}
\description{
Custom ggplot2 themes inspired by Minnesota state brand colors and styles.
}
\author{
\strong{Maintainer}: Kvale Dorfish \email{dorian.kvale@state.mn.us}

}
\keyword{internal}
