% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mn.R
\name{theme_mn}
\alias{theme_mn}
\title{Minnesota brand ggplot2 theme}
\usage{
theme_mn(
  font_size = 14,
  title_size = 22,
  banner = TRUE,
  banner_color = "blue",
  banner_text_color = "white",
  banner_alpha = 1,
  dark = FALSE,
  bg_color = NULL,
  text_color = NULL,
  hide_legend = TRUE,
  hide_legend_title = TRUE,
  legend_size = 15,
  hide_x_grid = FALSE,
  hide_y_grid = FALSE
)
}
\arguments{
\item{font_size}{Numeric. Base font size for plot text. Default is 14.}

\item{title_size}{Numeric. Font size for plot titles. Default is 22.}

\item{banner}{Logical. If TRUE, adds a colored title banner. Default is TRUE.}

\item{banner_color}{Character. Background color of the banner. Default is "blue".}

\item{banner_text_color}{Character. Text color of the banner. Default is "white".}

\item{banner_alpha}{Numeric from 0 to 1. Transparency of banner background. Default is 1.}

\item{dark}{Logical. If TRUE, applies a dark theme with customized background and text colors. Default is TRUE.}

\item{bg_color}{Character or NULL. Background color of entire plot when dark mode is enabled. Default is NULL (uses theme default).}

\item{text_color}{Character or NULL. Text color of plot when dark mode is enabled. Default is NULL (uses theme default).}

\item{hide_legend}{Logical. If TRUE, hides the plot legend. Default is TRUE.}

\item{hide_legend_title}{Logical. If TRUE, hides the legend title. Default is TRUE.}

\item{legend_size}{Numeric. Size of legend text. Default is 15.}

\item{hide_x_grid}{Logical. If TRUE, removes x-axis grid lines. Default is FALSE.}

\item{hide_y_grid}{Logical. If TRUE, removes y-axis grid lines. Default is FALSE.}
}
\value{
A `ggplot2` theme object added to a plot with ` + `.
}
\description{
Apply custom Minnesota-themed styling to a `ggplot2` plot. 
Allows for dark mode, a title banner, and legend and grid customization.
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(x = wt, y = mpg)) +
  geom_point(size = 3) +
  theme_mn()

}
