% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_banner.R
\name{title_banner}
\alias{title_banner}
\title{Add title banner for internal package use}
\usage{
title_banner(bg_color = "#003865", text_color = "#F0F0F0", bg_alpha = 1)
}
\description{
Add title banner for internal package use
}
\keyword{internal}
